update the node api

+++ sovrin_common/txn_util.py
@@ -116,7 +116,7 @@ def getTxnOrderedFields():
         (f.IDENTIFIER.nm, (str, str)),
         (f.REQ_ID.nm, (str, int)),
         (f.SIG.nm, (str, str)),
-        (TXN_TIME, (str, float)),
+        (TXN_TIME, (str, int)),
         (TXN_TYPE, (str, str)),
         (TARGET_NYM, (str, str)),
         (VERKEY, (str, str)),
